# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuickServe - Beverage stand order management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  caisse: 'caisse',
  preparateur: 'preparateur',
  admin: 'admin',
} as const;

export interface LoginInput {
  password: string;
  role: LoginInputRole;
  /** @nullable */
  eventSlug: string | null;
}

export interface LoginResult {
  success: boolean;
  role: string;
  /** @nullable */
  eventSlug?: string | null;
}

export interface SessionInfo {
  authenticated: boolean;
  /** @nullable */
  role: string | null;
  /** @nullable */
  eventSlug?: string | null;
}

export interface Event {
  id: number;
  nom: string;
  slug_url: string;
  actif: boolean;
  created_at: string;
}

export interface EventInput {
  nom: string;
  slug_url: string;
  actif?: boolean;
}

export interface EventUpdate {
  nom?: string;
  slug_url?: string;
  actif?: boolean;
}

export interface Article {
  id: number;
  evenement_id: number;
  nom: string;
  /** @nullable */
  description?: string | null;
  prix: number;
  /** @nullable */
  image_url?: string | null;
  stock_total: number;
  stock_disponible: number;
  disponible: boolean;
  created_at: string;
}

export interface ArticleInput {
  nom: string;
  description?: string;
  prix: number;
  image_url?: string;
  stock_total: number;
  disponible?: boolean;
}

export interface ArticleUpdate {
  nom?: string;
  description?: string;
  prix?: number;
  image_url?: string;
  stock_total?: number;
  disponible?: boolean;
}

export type OrderItemStatutLivraison = typeof OrderItemStatutLivraison[keyof typeof OrderItemStatutLivraison];


export const OrderItemStatutLivraison = {
  non_livre: 'non_livre',
  livre: 'livre',
} as const;

export interface OrderItem {
  id: number;
  commande_id: number;
  article_id: number;
  /** @nullable */
  article_nom?: string | null;
  quantite: number;
  prix_unitaire: number;
  statut_livraison: OrderItemStatutLivraison;
}

export type OrderStatut = typeof OrderStatut[keyof typeof OrderStatut];


export const OrderStatut = {
  en_attente: 'en_attente',
  reservee: 'reservee',
  payee: 'payee',
  livree_partiellement: 'livree_partiellement',
  livree: 'livree',
  expiree: 'expiree',
} as const;

export interface Order {
  id: number;
  evenement_id: number;
  nom_commande: string;
  statut: OrderStatut;
  montant_total: number;
  paye_cb?: number;
  paye_especes?: number;
  paye_cheque?: number;
  created_at: string;
  updated_at: string;
  /** @nullable */
  expiration_reservation?: string | null;
  items?: OrderItem[];
}

export interface OrderItemInput {
  article_id: number;
  quantite: number;
}

export interface OrderInput {
  nom_commande: string;
  items: OrderItemInput[];
}

export interface PaymentInput {
  paye_cb: number;
  paye_especes: number;
  paye_cheque: number;
}

export interface PartialDeliveryInput {
  item_ids: number[];
}

export interface Settings {
  id: number;
  evenement_id: number;
  temps_reservation_minutes: number;
  vente_ouverte: boolean;
  updated_at: string;
}

export interface SettingsUpdate {
  temps_reservation_minutes?: number;
  mdp_caisse?: string;
  mdp_preparateur?: string;
  mdp_admin?: string;
  vente_ouverte?: boolean;
}

export interface TopArticle {
  article_id: number;
  nom: string;
  total_vendu: number;
  chiffre_affaires: number;
}

export interface DashboardStats {
  ca_total: number;
  ca_cb: number;
  ca_especes: number;
  ca_cheque: number;
  nb_commandes_en_attente: number;
  nb_commandes_reservees: number;
  nb_commandes_payees: number;
  nb_commandes_livrees: number;
  nb_commandes_expirees: number;
  top_articles: TopArticle[];
}

export interface OrdersSummary {
  en_attente: number;
  reservee: number;
  payee: number;
  livree_partiellement: number;
  livree: number;
  expiree: number;
  livrees_aujourd_hui: number;
  articles_a_preparer?: number;
}