# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QuickServe - Beverage stand order management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with a role password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid password
...
  /events/{eventId}/articles:
    get:
      operationId: listArticles
      tags: [articles]
      summary: List articles for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Article"
    post:
      operationId: createArticle
      tags: [articles]
      summary: Create article for event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
  /articles/{id}:
    patch:
      operationId: updateArticle
      tags: [articles]
      summary: Update article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Article"
    delete:
      operationId: deleteArticle
      tags: [articles]
      summary: Delete article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /events/{eventId}/dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get dashboard stats for event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /events/{eventId}/orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [dashboard]
      summary: Get order counts by status
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersSummary"
components:
  schemas:
...
    LoginInput:
      type: object
      required: [password, role, eventSlug]
      properties:
        password:
          type: string
        role:
          type: string
          enum: [caisse, preparateur, admin]
        eventSlug:
          type: ["string", "null"]
    LoginResult:
      type: object
      required: [success, role]
      properties:
        success:
          type: boolean
        role:
          type: string
        eventSlug:
          type: ["string", "null"]
...
    Article:
      type: object
      required: [id, evenement_id, nom, prix, stock_total, stock_disponible, disponible, created_at]
      properties:
        id:
          type: integer
        evenement_id:
          type: integer
        nom:
          type: string
        description:
          type: ["string", "null"]
        prix:
          type: number
        image_url:
          type: ["string", "null"]
        stock_total:
          type: integer
        stock_disponible:
          type: integer
        disponible:
          type: boolean
        created_at:
          type: string
    ArticleInput:
      type: object
      required: [nom, prix, stock_total]
      properties:
        nom:
          type: string
        description:
          type: string
        prix:
          type: number
        image_url:
          type: string
        stock_total:
          type: integer
        disponible:
          type: boolean
    ArticleUpdate:
      type: object
      properties:
        nom:
          type: string
        description:
          type: string
        prix:
          type: number
        image_url:
          type: string
        stock_total:
          type: integer
        disponible:
          type: boolean
...
    TopArticle:
      type: object
      required: [article_id, nom, total_vendu, chiffre_affaires]
      properties:
        article_id:
          type: integer
        nom:
          type: string
        total_vendu:
          type: integer
        chiffre_affaires:
          type: number
    DashboardStats:
      type: object
      required: [ca_total, ca_cb, ca_especes, ca_cheque, nb_commandes_en_attente, nb_commandes_reservees, nb_commandes_payees, nb_commandes_livrees, nb_commandes_expirees, top_articles]
      properties:
        ca_total:
          type: number
        ca_cb:
          type: number
        ca_especes:
          type: number
        ca_cheque:
          type: number
        nb_commandes_en_attente:
          type: integer
        nb_commandes_reservees:
          type: integer
        nb_commandes_payees:
          type: integer
        nb_commandes_livrees:
          type: integer
        nb_commandes_expirees:
          type: integer
        top_articles:
          type: array
          items:
            $ref: "#/components/schemas/TopArticle"
    OrdersSummary:
      type: object
      required: [en_attente, reservee, payee, livree_partiellement, livree, expiree, livrees_aujourd_hui]
      properties:
        en_attente:
          type: integer
        reservee:
          type: integer
        payee:
          type: integer
        livree_partiellement:
          type: integer
        livree:
          type: integer
        expiree:
          type: integer
        livrees_aujourd_hui:
          type: integer
        articles_a_preparer:
          type: integer