# lib/api-zod/src/generated/types/articleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuickServe - Beverage stand order management API
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleUpdate {
  nom?: string;
  description?: string;
  prix?: number;
  image_url?: string;
  stock_total?: number;
  disponible?: boolean;
}