# lib/api-zod/src/generated/types/loginResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QuickServe - Beverage stand order management API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginResult {
  success: boolean;
  role: string;
  /** @nullable */
  eventSlug?: string | null;
}